chore(metrics): Moves scoring_metrics to metrics.
