+++ archai/nlp/metrics/text_predict/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+"""Text Predictor-related classes and methods.
+"""
