+++ archai/nlp/metrics/text_predict/wrappers/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+"""Wraps Archai functionalities to comply with Text Predict
+expected inputs and outputs.
+"""
