+++ archai/nlp/metrics/text_predict/wrappers/onnx_model_wrapper.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+
+"""ONNX-based models that works with the Text Predictor.
+"""
